+++ src/components/Cart/CartItem.js
@@ -2,12 +2,14 @@ import styled from "styled-components";
 import ItemCouter from "./ItemCounter";
 import { useState } from "react";
 import useCartStore from "../../hooks/useCartStore";
+import Item from "../Common/Item";
 
 const CartItem = ({
     name,
-    price,
     count = 0,
     storeName,
+    price,
+    desc,
     ImgSrc = "https://via.placeholder.com/100",
 }) => {
     const [quantity, setQuantity] = useState(count);
@@ -30,17 +32,9 @@ const CartItem = ({
         updateItemCount(storeName, name, 0);
     };
 
-    const formatPrice = (price) => {
-        return new Intl.NumberFormat("ko-KR").format(price) + "원";
-    };
-
     return (
         <Container>
-            <MenuImage src={ImgSrc} alt="메뉴 이미지" />
-            <ItemInfo>
-                <ItemTitle>{name}</ItemTitle>
-                <ItemPrice>{formatPrice(price)}</ItemPrice>
-            </ItemInfo>
+            <Item name={name} price={price} desc={desc} ImgSrc={ImgSrc} />
             <EraseButton onClick={handleRemove}>삭제</EraseButton>
             <ItemCouter
                 quantity={quantity}
@@ -67,26 +61,6 @@ const Container = styled.div`
     display: flex;
     gap: 20px;
     position: relative;
-`;
-
-const MenuImage = styled.img`
-    width: 64px;
-    height: 64px;
-`;
-
-const ItemInfo = styled.div`
-    display: flex;
-    flex-direction: column;
-`;
-
-const ItemTitle = styled.div`
-    margin-top: 5px;
-    font-size: 16px;
-    font-weight: bold;
-`;
-
-const ItemPrice = styled.div`
-    font-size: 14px;
 `;
 
 export default CartItem;
